Corregir timezone, decimales en grasa_visceral y mejoras en consultas
Timezone (fechas puras sin hora):
- Agregar timeZone:'UTC' en toLocaleDateString en pacientes, dashboard, ConsultaHistory
- calcularEdad: parsear fecha con split('-') para evitar desplazamiento UTC
- PacienteForm: usar .slice(0,10) en lugar de toISOString().split('T')[0]

grasa_visceral:
- Cambiar Int->Float en schema Prisma
- Quitar .int() y validación de entero en Zod y formularios
- parseInt->parseFloat en ConsultaForm

agregar-historica:
- Agregar validateField, fieldErrors, handleNumericChange, IMC preview
- Deshabilitar campo Próxima Cita Sugerida (no aplica a históricas)
- Corregir fecha usando T12:00:00.000Z

+++ src/app/api/consultas/[id]/route.ts
@@ -19,7 +19,7 @@ const consultaUpdateSchema = z.object({
   grasa_corporal: z.number().min(0).max(100).nullable().optional(),
   porcentaje_agua: z.number().min(0).max(100).nullable().optional(),
   masa_muscular_kg: z.number().min(0.5).max(400).nullable().optional(),
-  grasa_visceral: z.number().int().min(0).max(60).nullable().optional(),
+  grasa_visceral: z.number().min(0).max(60).nullable().optional(),
 
   // Perímetros (cm)
   brazo_relajado: z.number().min(5).max(160).nullable().optional(),
